requirements.txt
streamlit
playwright
python-dotenv